// apps/frontend2/src/api/models/updateLinkRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLinkRequestLabel } from './updateLinkRequestLabel';
import type { UpdateLinkRequestUrl } from './updateLinkRequestUrl';

export interface UpdateLinkRequest {
  label?: UpdateLinkRequestLabel;
  url?: UpdateLinkRequestUrl;
}